+++ src/console/controllers/PluginController.php
@@ -59,44 +59,25 @@ class PluginController extends Controller
      */
     public function actionInstall(?string $handle = null): int
     {
-        $pluginsService = Craft::$app->getPlugins();
-
         if ($handle === null) {
-            if (!$this->interactive) {
-                $this->stderr('A plugin handle must be specified.' . PHP_EOL, Console::FG_RED);
-                return ExitCode::UNSPECIFIED_ERROR;
-            }
-
-            $uninstalledPluginInfo = [];
-            foreach ($pluginsService->getAllPluginInfo() as $h => $info) {
-                if (!$info['isInstalled']) {
-                    $uninstalledPluginInfo[$h] = [
-                        [$h, 'format' => [Console::FG_YELLOW]],
-                        $info['name']
-                    ];
+            $handle = $this->_pluginPrompt(
+                'The following uninstalled plugins are present:',
+                'There aren’t any uninstalled plugins present.',
+                'Choose a plugin handle to install:',
+                function(array $info) {
+                    return !$info['isInstalled'];
                 }
+            );
+            if (is_int($handle)) {
+                return $handle;
             }
-
-            if (empty($uninstalledPluginInfo)) {
-                $this->stdout('There aren’t any uninstalled plugins present.' . PHP_EOL);
-                return ExitCode::OK;
-            }
-
-            $this->stdout('The following uninstalled plugins are present:' . PHP_EOL . PHP_EOL);
-            $this->table(['Handle', 'Name'], $uninstalledPluginInfo);
-
-            $handle = $this->prompt(PHP_EOL . 'Choose which plugin to install:', [
-                'validator' => function(string $input) use ($uninstalledPluginInfo) {
-                    return isset($uninstalledPluginInfo[$input]);
-                }
-            ]);
         }
 
         $this->stdout("*** installing {$handle}" . PHP_EOL, Console::FG_YELLOW);
         $start = microtime(true);
 
         try {
-            $success = $pluginsService->installPlugin($handle);
+            $success = Craft::$app->getPlugins()->installPlugin($handle);
         } catch (\Throwable $e) {
             $success = false;
         } finally {
@@ -114,11 +95,25 @@ class PluginController extends Controller
     /**
      * Uninstalls a plugin.
      *
-     * @param string $handle The plugin handle.
+     * @param string|null $handle The plugin handle.
      * @return int
      */
-    public function actionUninstall(string $handle): int
+    public function actionUninstall(?string $handle = null): int
     {
+        if ($handle === null) {
+            $handle = $this->_pluginPrompt(
+                'The following plugins plugins are installed and enabled:',
+                'There aren’t any installed and enabled plugins.',
+                'Choose a plugin handle to uninstall:',
+                function(array $info) {
+                    return $info['isInstalled'] && $info['isEnabled'];
+                }
+            );
+            if (is_int($handle)) {
+                return $handle;
+            }
+        }
+
         $this->stdout("*** uninstalling {$handle}" . PHP_EOL, Console::FG_YELLOW);
         $start = microtime(true);
 
@@ -145,11 +140,25 @@ class PluginController extends Controller
     /**
      * Enables a plugin.
      *
-     * @param string $handle The plugin handle.
+     * @param string|null $handle The plugin handle.
      * @return int
      */
-    public function actionEnable(string $handle): int
+    public function actionEnable(?string $handle = null): int
     {
+        if ($handle === null) {
+            $handle = $this->_pluginPrompt(
+                'The following plugins are disabled:',
+                'There aren’t any disabled plugins.',
+                'Choose a plugin handle to enable:',
+                function(array $info) {
+                    return $info['isInstalled'] && !$info['isEnabled'];
+                }
+            );
+            if (is_int($handle)) {
+                return $handle;
+            }
+        }
+
         $this->stdout("*** enabling {$handle}" . PHP_EOL, Console::FG_YELLOW);
         $start = microtime(true);
 
@@ -172,11 +181,25 @@ class PluginController extends Controller
     /**
      * Disables a plugin.
      *
-     * @param string $handle The plugin handle.
+     * @param string|null $handle The plugin handle.
      * @return int
      */
-    public function actionDisable(string $handle): int
+    public function actionDisable(?string $handle = null): int
     {
+        if ($handle === null) {
+            $handle = $this->_pluginPrompt(
+                'The following plugins are enabled:',
+                'There aren’t any enabled plugins.',
+                'Choose a plugin handle to disable:',
+                function(array $info) {
+                    return $info['isInstalled'] && $info['isEnabled'];
+                }
+            );
+            if (is_int($handle)) {
+                return $handle;
+            }
+        }
+
         $this->stdout("*** disabling {$handle}" . PHP_EOL, Console::FG_YELLOW);
         $start = microtime(true);
 
@@ -194,5 +217,47 @@ class PluginController extends Controller
         $time = sprintf('%.3f', microtime(true) - $start);
         $this->stdout("*** disabled {$handle} successfully (time: {$time}s)" . PHP_EOL . PHP_EOL, Console::FG_GREEN);
         return ExitCode::OK;
+    }
+
+    /**
+     * @param string $tableMessage
+     * @param string $noPlugins
+     * @param string $prompt
+     * @param callable|null $filterCallback
+     * @return string|int
+     */
+    private function _pluginPrompt(string $tableMessage, string $noPlugins, string $prompt, ?callable $filterCallback)
+    {
+        if (!$this->interactive) {
+            $this->stderr('A plugin handle must be specified.' . PHP_EOL, Console::FG_RED);
+            return ExitCode::UNSPECIFIED_ERROR;
+        }
+
+        $pluginInfo = Craft::$app->getPlugins()->getAllPluginInfo();
+        if ($filterCallback) {
+            $pluginInfo = array_filter($pluginInfo, $filterCallback);
+        }
+
+        foreach ($pluginInfo as $handle => $info) {
+            $uninstalledPluginInfo[$handle] = [
+                [$handle, 'format' => [Console::FG_YELLOW]],
+                $info['name']
+            ];
+        }
+
+        if (empty($uninstalledPluginInfo)) {
+            $this->stdout($noPlugins . PHP_EOL);
+            return ExitCode::OK;
+        }
+
+        $this->stdout($tableMessage . PHP_EOL . PHP_EOL);
+        $this->table(['Handle', 'Name'], $uninstalledPluginInfo);
+        $this->stdout(PHP_EOL);
+
+        return $this->prompt($prompt, [
+            'validator' => function(string $input) use ($uninstalledPluginInfo) {
+                return isset($uninstalledPluginInfo[$input]);
+            }
+        ]);
     }
 }